+++ src/main/java/com/kakaotech/team14backend/outer/member/service/MemberService.java
@@ -5,7 +5,7 @@ import com.kakaotech.team14backend.inner.member.model.Role;
 import com.kakaotech.team14backend.inner.member.model.Status;
 import com.kakaotech.team14backend.inner.member.repository.MemberRepository;
 import com.kakaotech.team14backend.point.infrastructure.PointRepository;
-import com.kakaotech.team14backend.inner.point.usecase.CreatePointUsecase;
+import com.kakaotech.team14backend.point.appliation.CreatePointUsecase;
 import com.kakaotech.team14backend.inner.post.repository.PostLikeCountRepository;
 
 import java.util.Optional;
