+++ src/main/java/com/kakaotech/team14backend/point/appliation/PointService.java
@@ -3,8 +3,6 @@ package com.kakaotech.team14backend.point.appliation;
 import com.kakaotech.team14backend.post.exception.PostNotFoundException;
 import com.kakaotech.team14backend.inner.member.model.Member;
 import com.kakaotech.team14backend.point.domain.UsePointDecider;
-import com.kakaotech.team14backend.inner.point.usecase.UsePointUsecase;
-import com.kakaotech.team14backend.inner.point.usecase.ValidatePointByPopularPostUsecase;
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import com.kakaotech.team14backend.inner.post.usecase.SetPostInstaCountUsecase;
