+++ src/main/java/com/kakaotech/team14backend/point/presentation/PointController.java
@@ -9,7 +9,7 @@ import com.kakaotech.team14backend.common.MessageCode;
 import com.kakaotech.team14backend.post.exception.PostNotFoundException;
 import com.kakaotech.team14backend.exception.UserNotAuthenticatedException;
 import com.kakaotech.team14backend.inner.member.model.Member;
-import com.kakaotech.team14backend.inner.point.usecase.UsePointUsecase;
+import com.kakaotech.team14backend.point.appliation.UsePointUsecase;
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import com.kakaotech.team14backend.inner.post.usecase.SetPostInstaCountUsecase;
