+++ src/main/java/com/kakaotech/team14backend/post/application/PostService.java
@@ -3,7 +3,7 @@ package com.kakaotech.team14backend.post.application;
 import com.kakaotech.team14backend.image.domain.Image;
 import com.kakaotech.team14backend.image.application.CreateImage;
 import com.kakaotech.team14backend.point.domain.GetPointPolicy;
-import com.kakaotech.team14backend.inner.point.usecase.GetPointUsecase;
+import com.kakaotech.team14backend.point.appliation.GetPointUsecase;
 import com.kakaotech.team14backend.inner.post.usecase.FindMyPostUsecase;
 import com.kakaotech.team14backend.inner.post.usecase.FindNonAuthPostListUsecase;
 import com.kakaotech.team14backend.inner.post.usecase.FindPersonalPostListUsecase;
